[transform] refactor so core transform code is in separate module; start implementing computed properties; center init rect svg at origin
